add old AflStatsStage back and rename it to StatsStage
